cicdc fix

+++ src/har_analyzer/utils/helpers.py
@@ -13,7 +13,7 @@ def get_memory_usage() -> float:
         Memory usage in megabytes
     """
     process = psutil.Process()
-    return process.memory_info().rss / 1024 / 1024
+    return float(process.memory_info().rss / 1024 / 1024)
 
 
 def format_bytes(bytes_value: Union[int, float]) -> str:
